COZMO-12145 Debug 'erase all needs' feature (#4596)
* COZMO-12145 Debug 'erase all needs' feature

In general this is a feature very similar to the debug 'erase all' tool, found
in the data persistence tab

- Add the new WipeRobotNeedsData message to the SDK message blacklist
- Handle new WipeRobotNeedsData in NeedsManager; erase the device storage
  for needs, and then erase the robot storage for needs; send ETG
  RestoreRobotStatus when done
- DataPersistancePrefab: link up the button
- DataPersistance.cs:  Handle the button, similar to how the reset all button is

* COZMO-12145 Debug 'erase all needs' feature Part 2

- Engine:  NeedsManager:  Pull some init stuff out of Init into a new InitInternal;
  call from Init, and also after resetting all needs data
- DataPersistenceManager:  Don't all either 'all' or 'needs' buttons to be pressed
  if EITHER is in process
- Also, fix subtle bug where if you pressed 'erase all' with no robot connected,
  you'd never be able to erase all in that session because IsResettingEverything
  would always be true
- After getting signal back from engine after erase, use those bools to take
  one of two paths, based on whether we reset needs only or not
- Also, don't call WipeRobotNeedsData when we're also calling WipeRobotGameData,
  because the latter includes the former, and we don't want to do both

* COZMO-12145 debug erase all needs fixes

- Define new WipeDeviceNeedsData CLAD message
- Add WipeDeviceNeedsData to SDK blacklist
- Have NeedsManager listen to that, and delete the save file on device
- For WipeRobotGameData, call InitInternal (this is when the 'erase everything'
  button is pressed)
- WipeRobotNeedsData no longer wipes device data
- Unity:  act accordingly

+++ unity/Cozmo/Assets/DevTools/DebugMenu/DebugMenuPanes/DataPersistencePane/DataPersistencePane.cs
@@ -10,6 +10,8 @@ namespace DataPersistence {
     private Button _ResetSaveDataButton;
     [SerializeField]
     private Button _ResetEverythingButton;
+    [SerializeField]
+    private Button _ResetNeedsButton;
 
     [SerializeField]
     private Button _StartNewSessionButton;
@@ -32,6 +34,7 @@ namespace DataPersistence {
     [SerializeField]
     private Text _LblStatus;
     private bool _IsResettingEverything = false;
+    private bool _IsResettingNeeds = false;
 
     private void Start() {
       _ResetSaveDataButton.onClick.AddListener(HandleResetSaveDataButtonClicked);
@@ -39,6 +42,7 @@ namespace DataPersistence {
 
       _SubmitSkillsButton.onClick.AddListener(SubmitSkillsButtonClicked);
       _ResetEverythingButton.onClick.AddListener(SubmitResetEverythingData);
+      _ResetNeedsButton.onClick.AddListener(SubmitResetNeedsData);
       _LblStatus.text = "";
       InitSkills();
 
@@ -62,6 +66,11 @@ namespace DataPersistence {
       Anki.Cozmo.ExternalInterface.BlockPoolResetMessage blockPoolResetMessage = new Anki.Cozmo.ExternalInterface.BlockPoolResetMessage();
       RobotEngineManager.Instance.Message.BlockPoolResetMessage = blockPoolResetMessage;
       RobotEngineManager.Instance.SendMessage();
+
+      // Delete the needs state data from device
+      if (RobotEngineManager.Instance.CurrentRobot != null) {
+        RobotEngineManager.Instance.CurrentRobot.WipeDeviceNeedsData();
+      }
     }
 
     private void StartNewSessionButtonClicked() {
@@ -98,31 +107,56 @@ namespace DataPersistence {
     }
 
     private void SubmitResetEverythingData() {
-      if (!_IsResettingEverything) {
-        _IsResettingEverything = true;
+      if (!_IsResettingEverything && !_IsResettingNeeds) {
         if (RobotEngineManager.Instance.CurrentRobot == null) {
           _LblStatus.text = "Error: Not connected to the robot!";
           _LblStatus.color = Color.red;
         }
         else {
+          _IsResettingEverything = true;
           _LblStatus.text = "THINKING. Stop touching things.";
           _LblStatus.color = Color.blue;
+          RobotEngineManager.Instance.CurrentRobot.WipeDeviceNeedsData();
           RobotEngineManager.Instance.CurrentRobot.WipeRobotGameData();
+        }
+      }
+    }
+
+    private void SubmitResetNeedsData() {
+      if (!_IsResettingEverything && !_IsResettingNeeds) {
+        if (RobotEngineManager.Instance.CurrentRobot == null) {
+          _LblStatus.text = "Error: Not connected to the robot!";
+          _LblStatus.color = Color.red;
+        }
+        else {
+          _IsResettingNeeds = true;
+          _LblStatus.text = "THINKING. Stop touching things.";
+          _LblStatus.color = Color.blue;
+          RobotEngineManager.Instance.CurrentRobot.WipeDeviceNeedsData();
+          RobotEngineManager.Instance.CurrentRobot.WipeRobotNeedsData();
         }
       }
     }
 
     private void HandleRestoreStatus(Anki.Cozmo.ExternalInterface.RestoreRobotStatus status) {
       if (status.isWipe) {
-        _IsResettingEverything = false;
         if (status.success) {
-          HandleResetSaveDataButtonClicked();
+          if (_IsResettingEverything) {
+            HandleResetSaveDataButtonClicked();
+          }
+          else {
+            if (NeedsConnectionManager.Instance != null) {
+              NeedsConnectionManager.Instance.ForceBoot();
+            }
+          }
           DebugMenuManager.Instance.CloseDebugMenuDialog();
         }
         else {
           _LblStatus.text = "Error: Robot Data clear failed!";
           _LblStatus.color = Color.red;
         }
+        _IsResettingEverything = false;
+        _IsResettingNeeds = false;
       }
     }
